+++ loader/debug_utils.h
@@ -24,7 +24,6 @@
 
 #pragma once
 
-#include "vk_loader_platform.h"
 #include "loader_common.h"
 
 // General utilities
