+++ loader/dirent_on_windows.h
@@ -10,6 +10,8 @@
 
 */
 
+#include "loader_common.h"
+
 #ifdef __cplusplus
 extern "C" {
 #endif
